// ademax/docsearch_frontend_proj_copy/angular-ocr/src/app/modules/model/user.model.ts
export class UserAuth {
  isvalid: boolean;
  canchange: boolean;
  candelete: boolean;
  canadd: boolean;
  cantk: boolean;
  cantkmeta: boolean;
  cantkcate: boolean;
  cantkname: boolean;
  ismanager: boolean;
  name: string;
  isnew: boolean;
  timeout: number;
  token: string;
  username: string;
}